+++ CHANGELOG.md
@@ -12,14 +12,18 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Add Slicing Aided Hyper Inference (SAHI) for object detection to improve small objects
   recall at inference.
 - Add TensorRT export for LTDETR models.
-- Add ONNX and TensorRT export for instance segmentation models.
-- Add ONNX and TensorRT export for panoptic segmentation models.
-- Add ONNX and TensorRT export for semantic segmentation models.
+- Add FP16/FP32 ONNX and TensorRT export for instance segmentation models.
+- Add FP16/FP32 ONNX and TensorRT export for panoptic segmentation models.
+- Add FP16/FP32 ONNX and TensorRT export for semantic segmentation models.
 - Add example jupyter notebooks for ONNX and TensorRT export.
 - Add tiny semantic segmentation models.
+- Add tiny instance segmentation models.
+- Add tiny panoptic segmentation models.
 - Add support for Ultralytics RT-DETR models.
 - PicoDet family of object detection models for low-power embedded devices.
 - Add support for different patch size in EoMT and semantic segmentation.
+- Add classwise metrics support for object detection models.
+- Add ignore_classes support for object detection models.
 
 ### Changed